requirements.txt
streamlit
google.generativeai
pillow
python-dotenv